Update pgcopydb sentinel's replay_lsn asynchronously.
When in follow replay mode, the pgcopydb sentinel is updated quite
frequently so that we can see progress being made and be reactive about when
to implement the migration cutover.

Now, we might have to consider some latency with the source database system
where the replay_lsn is published, so rate-limiting those sentinel updates
is important. Also, rate limiting alone might not be enough to prevent
slowing down the replay process, so switch to using libpq async API.