Introduce new pgcopydb config get|set commands.
Those commands only work when the config.ini file can be found on-disk,
which is to say when a pgcopydb command has been started already. The
commands allow to review and edit the current settings being used.

At the moment editing the values still has no impact on the running
processes.