Fix the transition between replay operating modes. (#277)
When switching replay operating modes we want to replay all the changes that
has been streamed locally but have not been caught-up yet. This includes
some JSON to SQL transformation work that was missed before.

To be able to receive from the transform queue after having received a
signal to stop, we need to also reset the signals control flags, otherwise
all we get is an early exit in queue_receive.