Fix how snapshot are exported when using CDC.
It turns out that we can't just call pg_create_logical_replication_slot() in
a transaction that uses a snapshot to sync the initial copy of the data with
the change data capture. We need to use the CREATE_REPLICATION_SLOT logical
replication command with the RESERVE_WAL true SNAPSHOT 'export' options.

For now this only adds a new test case that fails.