In --follow mode, export the snapshot via CREATE_REPLICATION_SLOT.
The previous code would set transaction snapshot to a previously exported
snapshot and then create the logical replication slot, but that creates a
replication gap. Postgres supports only one way to avoid that gap, which
consists of exporting the snapshot used to create the replication slot.

In passing, register the full information we get from creating the
replication slot to file, in such a way that we can re-use that file and
information later, including the consistent_point LSN that is associated
with the replication slot... and the snapshot.

As a result, the following commands have been removed from the CLI:

  pgcopydb stream create slot
  pgcopydb stream create origin
  pgcopydb stream drop slot
  pgcopydb stream drop origin

The only way to create a replication slot with pgcopydb is now the command
`pgcopydb snapshot --follow`, and the replication origin on the target must
be created with the consistent_point LSN returned by the same command.