Implement schema editing.
Add a facility to dump DDL commands for a specific set of tables, and
include also their indexes and the foreign key definitions that point to
selected tables. The command is pgcopydb dump sql-files --filters ...

Each SQL object definition is saved in their own file: <oid>.sql

Later implement a way to use the saved files instead of the pg_dump archive
contents when preparing the schema on the target database, in a way that if
the SQL files have been edited the new schema is installed there.