Fix empty xid and timestamp for continued txn COMMIT messages.
This fixes the missing `xid` and `timestamp` for continued transactions,
```sql
COMMIT; -- {"xid":0,"lsn":"6/A50001C8","timestamp":""}
```

This timestamp is used in `pgsql_replication_origin_xact_setup` which we
update along with `COMMIT;` statement. So it makes more sense to put
COMMIT's `timestamp` as origin timestamp.

Retaining the current update with `BEGIN` action as well. This avoids
`timestamp` being empty for `BEGIN` SQL statement in the file when COMMIT
spills to next SQL file.