Fix apply when file starts with non-begin/keepalive statements.
This fixes the issues where we skip applying I/U/D/T/C leading to the
warning "there is already a transaction in progress" when a txn continues
to next file and first statement of that file is not a BEGIN or KEEPALIVE
statement. In that case, `reachedStartPos` is never set to true and apply
never starts.

This happens because `reachedStartPos` is set to `false` everytime a new
file is processed.