Fix missing commit_lsn for continued transaction.
When a transaction is split into multiple JSON WAL files, the transformed
BEGIN statement does not include the commit_lsn since the COMMIT statement
appears in the next file. This poses a challenge during the apply process,
as the commit_lsn is necessary to determine when to start applying the
current and subsequent changes.

To resolve this issue, this fix parse the commit_lsn during the apply phase
from a transaction metadata file. This file is created during the transform
phase as soon as the COMMIT message is processed. However, this also
implies that the apply process must wait for the creation of this file
before it can begin applying the changes.

Fixes: #299