Use JSON format for work-directory summary files. (#300)
This ensures that Postgres object names are properly handled even when they
may contain multiple lines, which is accepted, possible... and then found in
the field.

In passing, ensure that we VACUUM ANALYZE tables without indexes. It's
almost relevant to that patch because the failure mode of multiple-lines
table names would break our vacuum module.