Fix file rotation issue during streaming. (#298)
* Fix file rotation issue during streaming.

Fixes a bug that occurs when multiple transactions have BEGIN and other
statements in one WAL file, while their COMMITs are in the next file. The
issue arises when pgcopydb receives the COMMIT of the first transaction,
causing it to move to the next file. However, upon receiving the BEGIN of
the next transaction, it incorrectly goes back to the previous file,
resulting in an inconsistent file rotation pattern.

To address this issue, this commit introduces tracking the LSN of the last
message written to the file. With this, file rotation is skipped if the
incoming message has an LSN less than the lastWrittenLSN, ensuring that
writes consistently occur in the current or next file.

* Rename lastWrite to lastWriteTime.

* Fix missing commit_lsn for continued transaction.

When a transaction is split into multiple JSON WAL files, the transformed
BEGIN statement does not include the commit_lsn since the COMMIT statement
appears in the next file. This poses a challenge during the apply process,
as the commit_lsn is necessary to determine when to start applying the
current and subsequent changes.

To resolve this issue, this fix parse the commit_lsn during the apply phase
from a transaction metadata file. This file is created during the transform
phase as soon as the COMMIT message is processed. However, this also
implies that the apply process must wait for the creation of this file
before it can begin applying the changes.

Fixes: #299