apply: Remove unnecessary reading for commit_lsn after reaching startpos.
Currently, we use txnCommitLSN to determine if we have reached the
startpos. If the txnCommitLSN is invalid, we read the transaction metadata
file from disk or wait until it exists. However, there is a bug that we
continue to check the validity of txnCommitLSN and read metadata file even
after reaching startpos. This additional work is unnecessary.

Furthermore, this commit partially address an issue where a transaction is
split by KEEPALIVE message during the replay phase. Previously, we would
wait for transaction metadata file to be created, which would never happen
and eventually lead to migration failure.