Refactor code to prepare for test_decoding improvements.
To finish implementing test_decoding support, we need to be able to look-up
attribute names for them being part of a primary-key definition, in order to
process UPDATE statements without the old-key: and new-key: parts.

In order to do that we prepare internal data structures to make it easy to
provide our internal catalogs representation to the transform sub-processes,
both when using the catchup mode API (from disk) or the replay mode
API (from a stream).

In passing, share more code in the ld_transform.c module in order to
simplify maintenance. It also looks like we're going to use much less memory
to implement the same job, that's a nice side effect of this refactoring.