Introduce a new pgcopydb internal message: ENDPOS.
When reaching endpos in the middle of a transaction we should stop
processing the stream, transform what we have, and apply up to the last full
transaction, ignoring the partial one at the end.

Because of the streaming approach taken by pgcopydb, the next best thing we
can do is to ROLLBACK the last partial transaction while still being able to
recognise that --endpos has been reached.

Previously to this patch, pgcopydb would use the endpos LSN to forge a
keepalive message, meaning that we would then update our replay_lsn and
replication origin tracking to a position in the middle of a transaction we
did not replay.

In case when the user then restart pgcopydb with a new endpos then we would
skip one transaction which was written off as already replayed. This patch
fixes this situation.