Improve parsing of UPDATE messages from test_decoding. (#329)
The test_decoding logical decoding output module sometimes issues UPDATE
statements with a single list of columns and values, and the client side is
then supposed to know what columns are in the WHERE clause and which in the
SET clause.

To know that, we need to have a version of the catalogs in-memory and
implement lookups. This PR does that.