Fix transforming UPDATE messages WHERE/SET clauses.
The logical decoding messages both from test_decoding and wal2json might
contain the same column in both WHERE and SET clauses with the same value.
When that's the case, it's best to reduce the UPDATE statement and refrain
from adding the same column-value pair in both clauses.

Also that's not just an optimisation now, as without this the following
could happen:

    ERROR:  column "id" can only be updated to DEFAULT
    DETAIL:  Column "id" is an identity column defined as GENERATED ALWAYS.