Fix resuming the transform process.
When the streaming is interrupted in the middle of a JSON/SQL file and then
resumes in replay mode, the transform process needs to read and parse the
latest JSON/SQL file in case it's being restarted in the middle of a
transaction.

The transform process rewrites the SQL file even when it was done
previously, as a way of implementing "cache invalidation" where the source
of truth is the JSON file. This PR implements the same approach to the
replay mode start-up sequence.