Fix stream_transform_resume to share context with stream_transform_file.
When resuming operations from pre-existing files on-disk, we implement some
cache invalidation by transforming the JSON file again from scratch, and
tracking progress in our internal data structures.

That's needed in case the previous version of the file terminates in the
middle of a transaction rather than at a clean transaction boundary.

The previous approach missed that the internal tracking state kept in the
StreamContext instance was not shared. Now it is.