Fix skipping streaming messages when resuming from latest JSON file.
It might happen that several JSON messages have the same LSN in our stream,
usually that concerns KEEPALIVE and BEGIN and COMMIT message. Unless paying
attention to the latest message action, we could skip a BEGIN message that
has the same LSN as the latest message even if that message was a KEEPALIVE.

Fix this by keeping around the latest message action. It should be all we
need to distinguish a new message from the latest one.