Add facilities to list extension versions.
The new commands allow listing extension versions in different ways and
formats (human-friendly table, JSON):

  $ pgcopydb list extensions --available-versions
  $ pgcopydb list extensions --requirements

The idea is to prepare for a feature where the extension versions to install
on the target database can be provided by the user. To prepare for that, the
new commands allow listing available extension versions on the target
database and preparing the current requirements file from the extensions
available on the target and their default version.