Assorted fixes for the test_decoding parser.
First, Postgres type names might contain square brackets (arrays), which are
used as a delimited in the test_decoding format for typenames. Fix the code
to handle the following properly:

    f2[text[]]:'{incididunt,ut,labore,et,dolore,magna}'

Then, it turns out that test_decoding emits text in standard SQL form, which
is to say in single-quotes and with single-quotes doubled, but with the \n
and other characters not-escaped. We need to edit the representation a
little so that the ld_transform module processes the text correctly.