Fix a couple bugs in pgcopydb clone --follow.
First, when interrupting the pgcopydb process interactively with C-c while
the transform process is running, we should exit the follow loop cleanup.

Second, when using pgcopydb clone --follow --restart we should do a
replication setup cleanup first, the same as the pgcopydb stream cleanup
command, so that we can actually restart from scratch as the command says.