Improve pgcopydb compare data checksum computation.
Use an output format that is stable in number of digits and can deal with a
bigint overflow (in Postgres a sum(bigint) is numeric): use an MD5 sum and
represent it as an UUID.

Also, include the row count in the MD5 computation to better protect against
collisions.

Finally, add support for pgcopydb compare data --json.