When applying --follow changes, set synchronous_commit to off. (#425)
This allows a way faster replay because we don't have to wait for Postgres
disk sync operation before sending the next SQL command. We need to be
careful with what replay_lsn value is sent back to the replication protocol,
and for that we need to introduce a tracking between the source LSN and the
replay insert LSN.

We still restart applying from the latest LSN that we manage to commit
durably using the replication origin API. The advantage of that API is that
it's as durable as the transactions replayed on the target system.