Fix error handling of fork() calls.
When failing to start a sub-process (out-of-memory can be reported when
calling the fork() system call), properly handle the error rather than
continue trying the migration.

In order to cancel the other sub-processes that were succesfully started we
then send a TERM signal to all processes in our process group.