Improve error handling when COPY fails.
In the normal case we want the COPY process to continue with the rest of the
tables in the queue, and in the fail-fast case we want to bubble-up the
problem as fast as possible for a quick total failure.

The partial failure in the normal case can be resumed with the existing
option such as --resume, as long as the snapshot still is held, or when
using the option --not-consistent.

Also, ensure that when errors happen mid-flight in the COPY processes we
still send the STOP messages in the queues, or signal all the other
processes to terminate.