Fix test runs for outside contributors
When an outside contributor submits a PR, the test runs fail because
they have the commits on a fork, and there is no easy way to get the
commits from the fork into the main repository. This change makes the
test runs work for outside contributors by always checking out the SHA
for the event.

We no longer need to have separate steps for PRs and non-PRs as
checkout@v3 has sane defaults for both. Below is an excerpt from the
docs of this action that explains the behavior of ref values:

  The branch, tag or SHA to checkout. When checking out the repository
  that triggered a workflow, this defaults to the reference or SHA for
  that event. Otherwise, uses the default branch.