Review logical decoding client tracking of LSNs.
When replaying from file, ensure that the last transaction of a file is
commited with synchronous_commit on, allowing better tracking of replication
progress.

Also, improve our way to keep track of the previousLSN and log when we reach
endpos (KEEPALIVE, COMMIT, ENDPOS) to help get a sense of what's happening.

By setting previousLSN to the proper value even when reaching endpos at a
KEEPALIVE message that we skip, we allow the overall system to exit properly
in case when there is no activity on the source server.