Enhance clean_results() to avoid blocking calls to PQgetResult().
In a debug session with gdb, now enabled in our docker-compose setup, it was
apparent that a process would block on a PQgetResult() call done in the
context of clear_results() when an error would happen during the logical
replication streaming.

Fix both the clear_results() behavior and also refrain from calling it on
the error path, just finish the connection without trying to get possibly
pending results that we're not going to need anyway.