Review process startup init and setting of endpos / --endpos.
When the command line option is used, we should not take the value that is
fetched from the source pgcopydb.sentinel table. Also protect some processes
against having already reached endpos at startup time.

In passing, try to test for Broken PIPE upon connection retries. Testing is
not conclusive on that parts, but it seems harmless.