Properly reset stdIn and stdOut booleans to false when needed.
Now that we have pro-active actions implemented to check if the PIPE might
be broken we need to pay attention to these variables.

Also in passing have the catchup process between follow loops happen in
sub-processes too. That way the process model stays the same even in that
area of the code: the actual processing is done by sub-processes.