Set application name to process title and pid (#553)
This change makes it easier to identify the connections of pgcopydb in
pg_stat_activity. The process title contains information on the current
operation, and the pid of the process is also included in the
application name. This way we can easily identify the connections of
pgcopydb in pg_stat_activity even if the process title is truncated.