Parallelize pg_restore operations (#561)
This commit adds a new option --restore-jobs to pgcopydb that allows
specifying how many jobs can be used to run pg_restore operations in
parallel. This option can also be set using the PGCOPYDB_RESTORE_JOBS
environment variable.

When this option is set to 1, pgcopydb will run pg_restore with the
--single-transaction option, and pgcopydb will behave the same as it
used to. Otherwise, pg_restore will be run with the --jobs option set
to the number of jobs specified by the user.

If the user does not supply the --restore-jobs option, or set the
environment variable, pgcopydb will use the index-jobs value as default.