Fix escaping of identifiers while transforming for wal2json plugin.
wal2json doesn't escape identifiers as test_decoding do. To fix this issue
we are using PQescapeIdentifier function. Even though this function doesn't
require PGconn to be connected, but creating an empty PGconn object was
unfeasible. Some fields of PGconn that are required by new versions of
libpq for PQescapeIdentifier are not available in older PGconn. So decided
to do this the right way, but creating PGconn with an actual server
connection. It turned out more simpler than managing a dummy PGconn object.