Fix error handling (--fail-fast) for copy workers.
When the COPY table-data workers fail before the iteration over the list of
tables is done, we have a race condition where the COPY supervisor would be
busy and is not using waitpid(), thus not reacting to worker process
termination.

To solve, limit the responsibilities of the COPY supervisor process even
more and create a new process that's responsible for iterating over the
table list and sending tables/parts to the queue, etc.